PeoplePage and MovieList 2 Wrappers swapped to 1 moved to genericStyles
