+++ src/common/tiles/generisStyles/styled.js
@@ -29,3 +29,13 @@ export const PeopleStyledLink = styled(Link)`
         transition: 0.3s;
     }
 `;
+
+export const Wrapper = styled.div`
+  max-width: 1368px;
+  margin: 0 auto;
+  background-color: ${({ theme }) => theme.colors.lightGrey};
+
+  @media (max-width: ${({ theme }) => theme.breakpoints.largest}) {
+    padding: 0 16px 16px 16px;
+  }
+`;
