+++ src/features/movies/MovieList/styled.js
@@ -1,15 +1,5 @@
 import styled from "styled-components";
 
-export const Wrapper = styled.div`
-  max-width: 1368px;
-  margin: 0 auto;
-  background-color: ${({ theme }) => theme.colors.lightGrey};
-
-  @media (max-width: ${({ theme }) => theme.breakpoints.largest}) {
-    padding: 0 16px 16px 16px;
-  }
-`;
-
 export const ContentContainer = styled.div`
   margin: 24px auto 0;
   display: grid;
